+++ bindings/dart/rust/src/api/mod.rs
@@ -17,3 +17,4 @@
 
 #![allow(unexpected_cfgs)]
 pub mod opendal_api;
+mod capability;
